Rename provider to gcp 1. Per review comments so secret and parameter manager both live in    single provider 2. Common code/scaffolding will be part of initial PR along with Secret    Manager provider.
