--- providers/gcp-secret-manager/CHANGELOG.md
@@ -1,7 +1,0 @@
-# Changelog
-
-## 0.0.1
-
-### ✨ New Features
-
-* Initial release of the GCP Secret Manager OpenFeature provider.
