+++ providers/gcp/CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+## 0.0.1
+
+### ✨ New Features
+
+* Initial release with scaffolding for Google Cloud and GCP Secret Manager OpenFeature provider.
