+++ providers/gcp/src/test/resources/fixtures/bool-flag.txt
