+++ providers/gcp/src/test/resources/fixtures/double-flag.txt
