+++ providers/gcp/src/test/resources/fixtures/int-flag.txt
