+++ providers/gcp/src/test/resources/fixtures/object-flag.json
