+++ providers/gcp/src/test/resources/fixtures/string-flag.txt
