+++ providers/gcp/src/test/resources/log4j2-test.xml
