+++ providers/gcp/version.txt
